change put method
