+++ routes/notes.js
@@ -1,28 +1,12 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const Joi = require('joi');
-
+const isDateValid = require('../validations/validationHandler');
 const jsonParser = bodyParser.json();
 const router = express.Router();
 
-//createNoteValidation
-const fetchedData = Joi.object({
-    id: Joi.string().required(),
-    title: Joi.string().min(3).required(),
-    content: Joi.string().max(500),
-    createDate: Joi.date(),
-    updateDate: Joi.date().allow(null),
-});
-
 let notes = [];
 let currentNote;
-const isDataValid = (req, res, next) => {
-    const { error } = fetchedData.validate(req.body);
-    if (error) {
-        return res.status(404).json({ error: { message: error.message } });
-    }
-    next();
-};
+
 const findElement = (req, res, next) => {
     try {
         currentNote = notes.find((note) => note.id === req.params.id);
@@ -36,7 +20,6 @@ const findElement = (req, res, next) => {
 };
 
 const putElement = (req, res, next) => {
-    // notes = [req.body];
     notes = notes.map((note) => (note.id === req.params.id ? req.body : note));
     next();
 };
@@ -45,7 +28,7 @@ router.get('/notes', (req, res) => {
     res.send(notes);
 });
 
-router.post('/notes', [jsonParser, isDataValid], (req, res) => {
+router.post('/notes', [jsonParser, isDateValid], (req, res) => {
     notes.push(req.body);
     res.json(req.body);
 });
